requirements.txt
numpy
pandas
pillow
matplotlib
pypi-json
tqdm
sklearn
torch
torchvision
PyYAML